+++ src/components/ChoosedDay/TasksCopmonents/TasksColumnsList/TasksColumnsList.jsx
@@ -1,5 +1,7 @@
 import { TasksColumn } from '../TasksColumn/TasksColumn';
 import { StretchContainer, TaskContainer } from './TasksColumnsListStyled';
+import { correctTitle } from '../helper/helper';
+import { toolTitle } from '../helper/helper';
 
 export const TasksColumnsList = ({ sortedTasksData }) => {
   const { done, inProgress, toDo } = sortedTasksData;
@@ -7,24 +9,8 @@ export const TasksColumnsList = ({ sortedTasksData }) => {
   //Ключи заголовков колонок
   const dataTitle = Object.keys(sortedTasksData);
   // console.log(dataTitle);
-
-  //Функция коррекции заголовков
-  const correctTitle = value => {
-    if (value === 'done') {
-      return value.charAt(0).toUpperCase() + value.slice(1);
-    } else {
-      const title = value.charAt(0).toUpperCase() + value.slice(1);
-      return title.substr(0, 2) + ' ' + title.substr(2, 8).toLowerCase();
-    }
-  };
-  //Функция коррекции для тулбра (запрос на бек)
-  const toolTitle = value => {
-    if (value === 'done') {
-      return value;
-    }
-    return value.substr(0, 2) + '-' + value.substr(2, 8).toLowerCase();
-  };
-  // console.log(toolTitle(dataTitle[0]));
+  console.log('DATATITLE--->', dataTitle);
+  console.log('DATATITLEFunc--->', correctTitle(dataTitle[2]));
 
   return (
     <TaskContainer>
